More Complex getter methods implemented
getString - returns the word or clue string
xPos, yPos - returns the respective coordinate
getLength - returns length of the word (a bit unneeded but we have the info handy)
getOrientation - returns the orientation of the word.
the enum Orientation is public within the InputTest class
can be accessed with values InputTest.Orientation.ACROSS
			    InputTest.Orientation.DOWN
feel free to shorten or rename if you need to.